Separando os arquivos do projeto e de testes para aprendizado
